Finishing touches for Scenario 2
