--- dotnet-core/scenario2/apiC/Startup.cs
@@ -1,31 +1,0 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace ApiC
-{
-  public class Startup
-  {
-    public Startup(IHostingEnvironment env)
-    {
-      var builder = new ConfigurationBuilder()
-          .SetBasePath(env.ContentRootPath)
-          .AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
-          .AddEnvironmentVariables();
-      Configuration = builder.Build();
-    }
-
-    public IConfigurationRoot Configuration { get; }
-
-    public void ConfigureServices(IServiceCollection services)
-    {
-      services.AddMvc();
-    }
-
-    public void Configure(IApplicationBuilder app)
-    {
-      app.UseMvc();
-    }
-  }
-}
